requirement.txt
pytorch==1.8.0
torchvision==0.9.0
torchaudio==0.8.0
cudatoolkit=11.1 
opencv-python
pillow
numpy
matplotlib
tensorboardX
scikit-learn
imageio
